// src/interfaces/global.ts
export type bidUser = {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  profilePicture: string | null;
  about: {
    address: string | null;
    bio: string | null;
    skills: string[] | null;
  };
  companyInfo: string | null;
};
// Types for the project data
export interface Project {
  id: string;
  userId: string;
  name: string;
  category?: string;
  skills: string[];
  budget: string;
  priceType: string;
  deadline: string;
  goal: string;
  scopeOfWork: string;
  experienceLevel: string;
  status: string;
  slug: string;
  createdAt: string;
  updatedAt: string;
  user?: {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    profilePicture?: string;
    about?: string | null;
    companyInfo?: string | null;
  };
  bid?: Array<{
    id: string;
    userId: string;
    projectId: string;
    amount: number;
    status: string;
    createdAt: string;
    updatedAt: string;
    bidUser: bidUser;
  }>;
  isFavorite: boolean;
  isBid: boolean;
}